feat: finished

+++ server/src/routes/memories.ts
@@ -4,8 +4,14 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook( "preHandler", async (request) => {
+    await request.jwtVerify()
+  })
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub
+      },
       orderBy: {
         createAt: "asc",
       },
@@ -18,7 +24,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
       };
     });
   });
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -28,7 +34,11 @@ export async function memoriesRoutes(app: FastifyInstance) {
       where: {
         id,
       },
-    });
+    })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
     return memory;
   });
 
@@ -44,13 +54,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "68940211-5cfd-4f29-beb1-beb2beb532c8",
+        userId: request.user.sub,
       },
     });
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -63,7 +73,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+     memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -76,11 +96,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
 
     await prisma.memory.delete({
       where: {
